// pc/src/services/apis/models/ProjectDcdxEconomicPerformanceStatsVo.ts
export interface ProjectDcdxEconomicPerformanceStatsVo {
    /**
     * 主键
     * @type {string}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    id?: string;
    /**
     * 市区名称
     * @type {string}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    district?: string;
    /**
     * 园区名称
     * @type {string}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    park?: string;
    /**
     * 年份
     * @type {number}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    year?: number;
    /**
     * 产值 - 实绩（万元）
     * @type {number}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    outputActual?: number;
    /**
     * 产值增长贡献 &#61; (项目形成产值 / 全市规上工业产值增量) * 100%（百分比，单位：%）
     * @type {number}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    outputGrowthContribution?: number;
    /**
     * 开票 - 实绩（万元）
     * @type {number}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    invoiceActual?: number;
    /**
     * 开票 - 预期（万元）
     * @type {number}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    invoiceExpected?: number;
    /**
     * 开票达效率 &#61; (开票实绩 / 预期开票) * 100%（小数形式）
     * @type {number}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    invoiceEfficiencyRate?: number;
    /**
     * 营收（万元）
     * @type {number}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    revenue?: number;
    /**
     * 利润（万元）
     * @type {number}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    profit?: number;
    /**
     * 利润率 &#61; (利润 / 营业收入) * 100%（小数形式）
     * @type {number}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    profitMargin?: number;
    /**
     * 税收 - 实绩（万元）
     * @type {number}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    taxActual?: number;
    /**
     * 税收 - 预期（万元）
     * @type {number}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    taxExpected?: number;
    /**
     * 税收达效率 &#61; (税收实绩 / 税收预期) * 100%（小数形式）
     * @type {number}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    taxEfficiencyRate?: number;
    /**
     * 所属产业链群，如：新能源、生物医药、高端装备等
     * @type {string}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    industrialChainCluster?: string;
    /**
     * 是否产业链群
     * @type {boolean}
     * @memberof ProjectDcdxEconomicPerformanceStatsVo
     */
    isLq?: boolean;
}